trans result type

+++ rpcman-demo-api/src/main/java/cn/ipman/rpcman/demo/api/UserService.java
@@ -6,9 +6,11 @@ public interface UserService {
 
     User findById(int id, String name);
 
-    int getId(int id);
+    long getId(long id);
 
     String getName(String name);
 
     String getName(int id);
+
+
 }
